backend/chat_assistant/requirements.txt
google-cloud-storage>=2.16.0
google-cloud-firestore>=2.15.0
google-cloud-vision>=3.5.4
google-cloud-translate>=3.12.1
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.7.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.2.0
firebase-admin>=6.5.0
google-adk>=0.4.3
functions-framework>=3.5.0
